Refactor and simplify build script
- Removed `BIN_PATH` variable from build.sh as it is no longer needed.
- Updated `build_matrixmath_project` function to remove the `namespace` variable.
- Removed `build_matrixmath_serializer_project` function from build.sh.
- Cleaned up bin directory using `clean_directory` function instead of `rm -rf`.
- Deleted `install_from_remote.sh` script as it is no longer required.
- Deleted `matrixmath_serializer.h` and associated serialization files from include and src directories.
- Renamed and relocated matrix and vector source files for better organization.
- Updated include paths in source files to reflect the new directory structure.
- Removed tests related to matrix serialization, and reorganized remaining test files for matrix and vector modules.
