+++ src/vector/vector_item.c
