+++ src/vector/vector_item.h
