+++ src/vector/vector_item_print.h
