--- tests/serialization/vector_serializer_tests.h
@@ -1,12 +1,0 @@
-#ifndef VECTOR_SERIALIZER_TESTS_H
-#define VECTOR_SERIALIZER_TESTS_H
-
-/**
- * Vectors serializer tests function.
- *
- * @return int
- *   The constant that represent the exit status.
- */
-int vector_serializer_tests();
-
-#endif
